- Updated zendesk.rb to send the tenant_id - Updated user_updater.rb to receive the tenant_id and use it to create the context - Updated DDL.sql with new data type for the Zendesk user_id - Updated README.md with new configuration details.

+++ db/ddl.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `zendesk_users` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `kb_account_id` varchar(255) NOT NULL,
-  `zd_user_id` int(11) NOT NULL,
+  `zd_user_id` bigint(11) NOT NULL,
   `created_at` datetime NOT NULL,
   `updated_at` datetime NOT NULL,
   PRIMARY KEY (`id`),
